Fixed missing icons from ungendered emoji list
It is not correct to remove all ^(Man|Woman) emoji from the list,
as some emoji are only available for one or the other or just not
in a generic "person" form.

The big-o for this is abysmal, but it's a code generator that is
literally only run whenever a new UTR spec comes out, so who cares?