Upgrade to Unicode 13
This update regenerates the source files from the Unicode 13 TR. Segoe
UI Emoji is upgraded to that shipping w/ Windows 11 21H2 (10.0.22000.194),
and the list of basic emoji has been updated to those supported by it.

Note that the names, groups, subgroups, and relative sort order of some
emoji have been changed to match the changes in the UTR. This is not
considered to be a breaking change as the individual emoji supported by
the library are not considered to be part of the semantic versioning
contract.